Created a new file and created a new exception handling
Created a new file called "What you need" which informs the user what they need downloaded. Created a new exception handling for link_validation because I was having errors when the link could not be split. Added the test for the exception in the tests file along with adding more documentation for the test file.